+++ pennylane_lightning/core/simulators/lightning_gpu/catalyst/LightningGPUSimulator.cpp
@@ -180,6 +180,20 @@ void LightningGPUSimulator::NamedOperation(
     auto &&dev_wires = getDeviceWires(wires);
     auto &&dev_controlled_wires = getDeviceWires(controlled_wires);
 
+    if (name == "PauliRot") {
+        RT_FAIL_IF(optional_params.size() != 1,
+                   "PauliRot operation requires one string "
+                   "parameter for the Pauli word");
+        RT_FAIL_IF(!controlled_wires.empty(),
+                   "Controlled PauliRot is not supported");
+        RT_FAIL_IF(this->tape_recording,
+                   "PauliRot operation is not supported when tape "
+                   "recording is active"); // TODO: support caching
+        this->device_sv->applyPauliRot(dev_wires, inverse, params,
+                                       optional_params[0]);
+        return;
+    }
+
     // Update the state-vector
     if (controlled_wires.empty()) {
         this->device_sv->applyOperation(name, dev_wires, inverse, params);