+++ pennylane_lightning/core/simulators/lightning_qubit/catalyst/LightningSimulator.hpp
@@ -133,14 +133,12 @@ class LightningSimulator final : public Catalyst::Runtime::QuantumDevice {
                        std::vector<QubitIdType> &wires) override;
     [[nodiscard]] auto GetDeviceShots() const -> std::size_t override;
 
-    void NamedOperation(const std::string &name,
-                        const std::vector<double> &params,
-                        const std::vector<QubitIdType> &wires,
-                        bool inverse = false,
-                        const std::vector<QubitIdType> &controlled_wires = {},
-                        const std::vector<bool> &controlled_values = {},
-                        [[maybe_unused]] const std::vector<std::string>
-                            &optional_params = {}) override;
+    void NamedOperation(
+        const std::string &name, const std::vector<double> &params,
+        const std::vector<QubitIdType> &wires, bool inverse = false,
+        const std::vector<QubitIdType> &controlled_wires = {},
+        const std::vector<bool> &controlled_values = {},
+        const std::vector<std::string> &optional_params = {}) override;
     using Catalyst::Runtime::QuantumDevice::MatrixOperation;
     void
     MatrixOperation(const std::vector<std::complex<double>> &matrix,