+++ .github/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added support for `qml.pauli_measure` simulation for Catalyst.
   [(#1327)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1327)
 
+- Added support for `qml.PauliRot` simulation for Catalyst.
+  [(#1322)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1322)
 
 <h3>Improvements 🛠</h3>